fix spreading of props & fix pagination on first query
The first query-params call was missing the skip to account for perPage skip so it was always showing the first three items instead of the first three items after pageX was skipped.

I spread queryParams instead of setting query-params=queryPrams as the later was sending the queryParams prop with its own nested queryParams.queryParams.