Use new component and add support when only a single rail is passed in
Update the section feed wrapper to use the new section feed conponent.  This also add support for when you only pass a single rail to display all section feed component in one row with a single rail.  Previousely it would force two rows with the content and right rail being split.  This does not support section with lengthy right rail.